Resolve config and registration conflicts while retaining both model sets. Move the Thinker graph test to the split build_graph suite introduced on main.

+++ src/mobius/tasks/_controlnet.py
@@ -36,7 +36,7 @@ class ControlNetTask(ModelTask):
             dtype=ir.DataType.FLOAT,
             shape=["batch", config.in_channels, "height", "width"],
         )
-        timestep = builder.input("timestep", dtype=ir.DataType.INT64, shape=["batch"])
+        timestep = builder.input("timestep", dtype=ir.DataType.FLOAT, shape=["batch"])
         encoder_hidden_states = builder.input(
             "encoder_hidden_states",
             dtype=ir.DataType.FLOAT,
